remove obsolete code, convert chat user admin credit to console form action

+++ src/wbs/console/forms/TextFormFieldSpec.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 import lombok.experimental.Accessors;
 
 import wbs.console.module.ConsoleModuleData;
+
 import wbs.framework.component.annotations.PrototypeComponent;
 import wbs.framework.data.annotations.DataAttribute;
 import wbs.framework.data.annotations.DataClass;
@@ -42,10 +43,10 @@ class TextFormFieldSpec {
 	Boolean nullable;
 
 	@DataAttribute
-	Integer minimumLength;
+	Long minimumLength;
 
 	@DataAttribute
-	Integer maximumLength;
+	Long maximumLength;
 
 	@DataAttribute
 	String viewPriv;
